[STYLE] move IChat to type.tsx from ChatWindow.tsx
