+++ frontend/components/main/chat_window/ChatField/NonDmChat.tsx
@@ -2,7 +2,7 @@
 
 import { Box, Typography } from "@mui/material";
 import { Dispatch, SetStateAction } from "react";
-import {IChat} from "../ChatWindow"
+import { IChat } from "@/type/type";
 
 interface Props {
   msgs: IChat[];
